fix(integrations): converte progressPercent de clamp para double

+++ lib/features/integrations/domain/services/platform_goal_progress_calculator.dart
@@ -50,7 +50,7 @@ abstract final class PlatformGoalProgressCalculator {
             platform: platform,
             targetAmount: target,
             actualAmount: actual,
-            progressPercent: percent.clamp(0, 999),
+            progressPercent: percent.clamp(0, 999).toDouble(),
             sharePercent: share * 100,
           ),
         );
